Move SovereignCosmosComponents to sovereign-chain-components
