+++ crates/sovereign/sovereign-chain-components/src/cosmos/mod.rs
@@ -3,4 +3,5 @@
    with the counterparty being a Sovereign rollup
 */
 
+pub mod components;
 pub mod impls;
