+++ crates/sovereign/sovereign-relayer/src/exts/cosmos/chain.rs
@@ -1,47 +1,25 @@
 use cgp_core::prelude::*;
 use hermes_cosmos_chain_components::components::delegate::DelegateCosmosChainComponents;
 use hermes_cosmos_relayer::contexts::chain::CosmosChain;
-use hermes_relayer_components::chain::impls::queries::query_and_convert_client_state::QueryAndConvertRawClientState;
-use hermes_relayer_components::chain::impls::queries::query_and_convert_consensus_state::QueryAndConvertRawConsensusState;
+use hermes_relayer_components::chain::traits::message_builders::connection_handshake::CanBuildConnectionHandshakeMessages;
+use hermes_relayer_components::chain::traits::message_builders::create_client::CanBuildCreateClientMessage;
+use hermes_relayer_components::chain::traits::message_builders::update_client::CanBuildUpdateClientMessage;
 use hermes_relayer_components::chain::traits::payload_builders::create_client::CanBuildCreateClientPayload;
 use hermes_relayer_components::chain::traits::payload_builders::update_client::CanBuildUpdateClientPayload;
-use hermes_relayer_components::chain::traits::queries::consensus_state::{CanQueryConsensusState, ConsensusStateQuerierComponent};
-use hermes_relayer_components::chain::traits::queries::consensus_state_height::{CanQueryConsensusStateHeight, CanQueryConsensusStateHeights, ConsensusStateHeightsQuerierComponent};
+use hermes_relayer_components::chain::traits::queries::client_state::CanQueryClientState;
+use hermes_relayer_components::chain::traits::queries::consensus_state::CanQueryConsensusState;
+use hermes_relayer_components::chain::traits::queries::consensus_state_height::{
+    CanQueryConsensusStateHeight, CanQueryConsensusStateHeights,
+};
 use hermes_relayer_components::chain::traits::types::client_state::HasClientStateFields;
 use hermes_relayer_components::chain::traits::types::create_client::HasCreateClientOptionsType;
-use hermes_sovereign_chain_components::cosmos::impls::queries::consensus_state_heights::QuerySovereignConsensusStateHeightsFromGrpc;
-use hermes_sovereign_chain_components::cosmos::impls::sovereign_to_cosmos::client::create_client_message::BuildCreateSovereignClientMessageOnCosmos;
-use hermes_sovereign_chain_components::cosmos::impls::sovereign_to_cosmos::client::update_client_message::BuildUpdateSovereignClientMessageOnCosmos;
-use hermes_sovereign_chain_components::cosmos::impls::sovereign_to_cosmos::connection_handshake_message::BuildSovereignConnectionHandshakeMessageOnCosmos;
-use hermes_relayer_components::chain::traits::message_builders::connection_handshake::{CanBuildConnectionHandshakeMessages, ConnectionHandshakeMessageBuilderComponent};
-use hermes_relayer_components::chain::traits::message_builders::create_client::{CanBuildCreateClientMessage, CreateClientMessageBuilderComponent};
-use hermes_relayer_components::chain::traits::message_builders::update_client::{CanBuildUpdateClientMessage, UpdateClientMessageBuilderComponent};
-use hermes_relayer_components::chain::traits::queries::client_state::{CanQueryClientState, ClientStateQuerierComponent};
+use hermes_sovereign_chain_components::cosmos::components::SovereignCosmosComponents;
 
 use crate::contexts::sovereign_chain::SovereignChain;
-
-pub struct SovereignCosmosComponents;
 
 delegate_components! {
     DelegateCosmosChainComponents {
         SovereignChain: SovereignCosmosComponents,
-    }
-}
-
-delegate_components! {
-    SovereignCosmosComponents {
-        ClientStateQuerierComponent:
-            QueryAndConvertRawClientState,
-        ConsensusStateQuerierComponent:
-            QueryAndConvertRawConsensusState,
-        UpdateClientMessageBuilderComponent:
-            BuildUpdateSovereignClientMessageOnCosmos,
-        CreateClientMessageBuilderComponent:
-            BuildCreateSovereignClientMessageOnCosmos,
-        ConnectionHandshakeMessageBuilderComponent:
-            BuildSovereignConnectionHandshakeMessageOnCosmos,
-        ConsensusStateHeightsQuerierComponent:
-            QuerySovereignConsensusStateHeightsFromGrpc,
     }
 }
